// src/shared/api/generated/schemas/auctionStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Logistics Auctions API
 * OpenAPI spec version: 0.0.1
 */

export type AuctionStatus = (typeof AuctionStatus)[keyof typeof AuctionStatus];

export const AuctionStatus = {
  active: "active",
  completed: "completed",
  cancelled: "cancelled",
  draft: "draft",
} as const;