// src/shared/api/generated/schemas/userTradingStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Logistics Auctions API
 * OpenAPI spec version: 0.0.1
 */

export type UserTradingStatus =
  (typeof UserTradingStatus)[keyof typeof UserTradingStatus];

export const UserTradingStatus = {
  Leading: "Leading",
  Losing: "Losing",
  Winner: "Winner",
  OutOfTrade: "OutOfTrade",
  NotParticipant: "NotParticipant",
} as const;